Add support for ssh tunnelling in cylc message
Check for the existance of the variable 'CYLC_SSH_TUNNEL_HOST' in the job
submit script. If it exists ssh messages will pass through this host first
before going to their final destination, i.e. instead of

    ssh $USER@$TARGET COMMAND

it will call

    ssh $CYLC_SSH_TUNNEL_HOST ssh $USER@TARGET COMMAND

The intended use case for this is on a HPC system where connections from
compute nodes to the outside world has to pass through the login node first. As
an example at NCI we have the following systems:

    accessdev - Users access cylc here
    vayu - HPC login node, jobs are submitted here to be run on the compute node
    v0000 - HPC compute node

The compute nodes are forbidden from talking to accessdev (in fact they can't
even see it on the network, so the is_remote_host function fails).